Refactor cartService and product Recommendation
