+++ src/app/components/product-recommend/product-recommend.component.ts
@@ -1,19 +1,15 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { ProductImage } from 'src/app/models/product';
-import { CartService } from 'src/app/services/cart.service';
+import { Component, OnInit } from '@angular/core';
+
+import { ProductRecommendationService } from './product-recommend.service';
 
 @Component({
   selector: 'app-product-recommend',
   templateUrl: './product-recommend.component.html',
   styleUrls: ['./product-recommend.component.scss'],
+  providers: [ProductRecommendationService],
 })
 export class ProductRecommendComponent implements OnInit {
-  @Input() imageArray: ProductImage[];
-  public dataService: CartService;
-
-  constructor(private service: CartService) {
-    this.dataService = this.service;
-  }
+  constructor(public service: ProductRecommendationService) {}
 
   ngOnInit() {}
 }
